% Generated by roxygen2 (4.0.2): do not edit by hand
\name{pupilPlot}
\alias{pupilPlot}
\title{pupilPlot}
\usage{
pupilPlot(data = NULL, variableNames = c(),
  excludeFromSelectors = c("SubjectNr"), trialTimeVariable = "TrialTime",
  pupilSizeVariables = list(`Baseline Corrected` = "size"),
  samplingFrequency = 60, trialDuration = 5000,
  subjectVariable = "SubjectNr", trialVariable = "Trial",
  subjectMeansInterval = c(0, trialDuration))
}
\arguments{
\item{data}{A full data.table with all relevant variables.}

\item{variableNames}{Names of variables in the experiment that you would like to plot by.}

\item{excludeFromSelectors}{Names of variables not to be used in selectors (e.g.: Subject numbers)}

\item{trialTimeVariable}{Name of variable holding the trial-relative timepoints.}

\item{pupilSizeVariables}{List of variable desriptions and names}

\item{samplingFrequency}{How many samples per second.}

\item{trialDuration}{How long did each trial last.}

\item{subjectVariable}{Name of variable holding subject numbers.}

\item{trialVariable}{Name of variable holding trial numbers.}

\item{subjectMeansInterval}{Interval for exporting subject means.}
}
\description{
This function starts a shiny server for plotting pupil data.
}
\examples{
Start the shiny server.
##not run
variables = c("Stim1","Stim2","SubjectNr","Experiment","ACC")
pupilPlot(alldatas,variables,trialDuration=7000,subjectVariable="SubjectNr")
}

